GRN UPDATE
Done:
Retirved Approved PO using PO number into GRN to make GRN for goods recieved.

Next:
Adding item to GRN List. There might be instances where an item ordered was not supplied and also an item supplied was not ordered, both scenarios should have a solution. In GRN there should be an option to remove an item from the list and also add an item which was not ordered but supplied.