Refactor database function calls to instance methods
This commit refactors the code in several files to change usage of new db_handler() to an instance method call using $db. It replaces the instantiations of db_handler and the calling of its methods directly. Instead, the database handler object is now stored in a variable $db, and this variable is used to call the methods.

This was done for better readability, maintainability, and to avoid unnecessary instantiation of the database handler. No functional changes are made by this commit.